BT: broadcom: Support split write sequence of Oreo HW IFace.
In oreo, the HW interface module has been introduced also known
as project treble. Thereby, some of the lower-level bt stack
has been moved/rewritten such that BT packets are now written
in a different way to the character devices corresponding to
the BT driver:
- First, the packet type is written.
- Second, the paket's data is written.

The current version of the driver assumes that the packets are
written in a single operation. This patch changes the write node
of the broadcom BT/FM/ANT driver to match the initially mentioned
sequence of passing packets.

Moreover, this patch makes enabling the HCI filter dependent on
the reception of a HCI RESET command such that multiple BT init
sequences can be handled. In Oreo, not only libbt-vendor but also
a controller module from "system/bt" performs such a sequence.
Several commands of the init sequence must be intercepted, since
the driver already initializes the chip and it would be rendered
non-functional otherwise.

Change-Id: I30f0f7f15f72c283c36a619df44e4aa6686117db
Signed-off-by: Alexander Diewald <Diewi@diewald-net.com>